+++ cr_zkteco_biometric_integration/wizard/__init__.py
@@ -8,3 +8,5 @@ from . import calculate_attendance_wizard
 from . import daily_summary_report_wizard
 from . import absence_report_wizard
 from . import daily_attendance_report_wizard
+from . import biometric_download_log_wizard
+